Add files via upload
Two versions of the final iteration of the potential were trained, the first by filtering out structures with DFT energies greater than 1 eV/atom, and the second using the full reference dataset (see Supplementary Information of the paper for the full protocol). 

The first version was used to generate all the quantitative data presented, as it achieved good accuracy on the structures relevant our study, with an energy RMSE of 13.9 meV/atom relative to DFT results at the PBEsol level of theory. The second version, which includes higher-energy dimers and random structures, although less accurate (energy RMSE of 23.1 meV/atom), does not fail when handling structures with closer contacts between atoms, and therefore it was used to relax the polycrystalline structures with randomly-oriented grains. (We consider a simulation to have `failed' if, during the MD run, atoms come closer to each other than 1 Å, collapsing the structure, or if atoms are lost during the simulation.)