Add files via upload
The training and test datasets used for fitting the two versions of the potential. The filtered dataset for the low-energy version of the potential only includes structures with calculated energies below 1 meV/atom (PBEsol) and, while the full dataset includes the higher energy structures as well (an extra 17 close-contact dimer structures and 112 RSS structures).