Add files via upload
The training and test datasets for the two versions of the potential. The filtered dataset contains only structures with calculated energy values below 1 meV/atom (PBEsol), while the full dataset also includes these higher energy structures (an extra 17 close-contact dimer structures and 112 RSS structures).
The code for creating the incremental UMAP plots presented in Fig.1 (with the settings used), together with a Jupyter Notebook which allows the user to view the dataset interactively, hovering over datapoints for more information, are additionally included.